Add a few explicit conversions to avoid warnings with VS
The Visual Studio compiler warns about conversions with loss of precision.
This fixes a few that cause major output (although there's still quite
a few to do).

+++ src/include/stir/scatter/ScatterSimulation.h
@@ -116,7 +116,7 @@ public:
     get_output_proj_data_sptr() const;
 
     inline int get_num_scatter_points() const
-    { return this->scatt_points_vector.size();}
+    { return static_cast<int>(this->scatt_points_vector.size());}
     //! Get the template ProjDataInfo
     shared_ptr<const ProjDataInfoCylindricalNoArcCorr> get_template_proj_data_info_sptr() const;
     //! Get the ExamInfo
